+++ lib/unit_of_measure.cs
@@ -1,4 +1,4 @@
 public interface UnitOfMeasure
 {
-  double convert(double amount, UnitOfMeasure unit);
+  decimal convert(decimal amount, UnitOfMeasure unit);
 }
